Let context words to be synsets
Allow contexts with words like this:

00001740-n#n#w1#2#1.0

which are not words but synsets. Later, when applying calculate_mcr_hr
family of functions, PPV attached to the synsets are 'lighten' with
the appropiate weight.

Note that the 3rd field, which was previously used to decide wether a
word is distinguished, has now a value of '2', stating that the cword
is indeed a synset. Synset words can't be distinguished.