cword: introduce the CWord::cwdistnolight type.
These are words in the context which has to be disambiguated, but are not
used when initializing the graph PV. They are useful when the weights of the
word's concepts are known a-priori. For instance, say 'car' has 3 concepts
and we know this weights:

car_1: 0.2  car_2:0.5 car_3:0.3

we can create a context like:

car_1###2#0.2 car_2###2#0.5 car_3###2#0.3 car##w1#3

the last 'cwdistnolight' type is needed to produce the output for 'car', but
we do not need to further light its associated concepts, as they already do
have a weight.